# Reduces allocation in attributes