@pellared 2024-06-26
Can you please add benchstat results to the PR description?