@Succo 2024-06-26
Hello 

Thanks for the interest.

Here is benchstat results

```goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/internal/attribute
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
             │   old.txt   │               new.txt               │
             │   sec/op    │   sec/op     vs base                │
SliceValue-8   166.1n ± 5%   144.7n ± 5%  -12.88% (p=0.000 n=10)

             │  old.txt   │              new.txt               │
             │    B/op    │    B/op     vs base                │
SliceValue-8   152.0 ± 0%   128.0 ± 0%  -15.79% (p=0.000 n=10)

             │  old.txt   │              new.txt               │
             │ allocs/op  │ allocs/op   vs base                │
SliceValue-8   3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)```

And the command used for each run
`go test -bench=BenchmarkSliceValue -benchtime=10s -count=10 -benchmem`