@Succo 2024-06-26
I'm open to removing the comparison in the bench and adding a test on the number of allocation, if you are aligned on it.