@Succo 2024-06-27
Hello

Here is the new benchstat result with all functions

```
goos: linux
goarch: amd64
pkg: go.opentelemetry.io/otel/internal/attribute
cpu: 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
                    │   old.txt   │               new.txt                │
                    │   sec/op    │    sec/op     vs base                │
BoolSliceValue-8      132.7n ± 9%   104.4n ±  9%  -21.30% (p=0.000 n=10)
Int64SliceValue-8     147.5n ± 4%   125.8n ± 11%  -14.72% (p=0.000 n=10)
Float64SliceValue-8   147.1n ± 4%   128.7n ± 83%  -12.54% (p=0.023 n=10)
StringSliceValue-8    183.1n ± 9%   153.5n ±  9%  -16.16% (p=0.000 n=10)
geomean               151.5n        126.9n        -16.24%

                    │   old.txt   │              new.txt               │
                    │    B/op     │    B/op     vs base                │
BoolSliceValue-8      32.000 ± 0%   8.000 ± 0%  -75.00% (p=0.000 n=10)
Int64SliceValue-8      88.00 ± 0%   64.00 ± 0%  -27.27% (p=0.000 n=10)
Float64SliceValue-8    88.00 ± 0%   64.00 ± 0%  -27.27% (p=0.000 n=10)
StringSliceValue-8     152.0 ± 0%   128.0 ± 0%  -15.79% (p=0.000 n=10)
geomean                78.34        45.25       -42.23%

                    │  old.txt   │              new.txt               │
                    │ allocs/op  │ allocs/op   vs base                │
BoolSliceValue-8      3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
Int64SliceValue-8     3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
Float64SliceValue-8   3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
StringSliceValue-8    3.000 ± 0%   2.000 ± 0%  -33.33% (p=0.000 n=10)
geomean               3.000        2.000       -33.33%
```

Allocation is always the same which does not surprise me.