@pellared 2024-06-27
Can you please add a changelog entry?