@Succo 2024-06-27
Added the Changelog, and a few modification to the benchmarks proposed by @MrAlias